requirements.txt
alembic==1.16.4
blinker==1.9.0
click==8.2.1
dominate==2.9.1
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
visitor==0.1.3
Werkzeug==3.1.3
